PMG-292 When pressing on the Dashboard icon - the Map Icon appears selected Moved dashboard component to core module
