+++ console/src/app/core/dashboard/dashboard.component.html
